Unify ip parsers
