+++ core/src/main/scala/org/http4s/parser/Rfc2616BasicRules.scala
@@ -21,7 +21,6 @@ import scala.reflect.ClassTag
 import scalaz.Validation
 import org.http4s.{ParseFailure, ParseResult}
 import org.http4s.internal.parboiled2._
-import org.http4s.internal.parboiled2.CharPredicate.{ HexDigit => HEXDIG }
 
 // direct implementation of http://www.w3.org/Protocols/rfc2616/rfc2616-sec2.html#sec2
 private[http4s] trait Rfc2616BasicRules extends Parser {
@@ -76,64 +75,6 @@ private[http4s] trait Rfc2616BasicRules extends Parser {
   def ListSep = rule { oneOrMore("," ~ OptWS) }
 
   def LASTCHAR: Rule1[Char] = rule { push(input.charAt(cursor - 1)) }
-
-  private val DIGIT = CharPredicate.Digit
-  private val DIGIT04 = CharPredicate('0' to '4')
-  private val DIGIT05 = CharPredicate('0' to '5')
-  private val DIGIT19 = CharPredicate.Digit19
-
-  def IPNumber = rule {
-    capture(
-      '2' ~ (DIGIT04 ~ DIGIT | '5' ~ DIGIT05)
-        | '1' ~ DIGIT ~ DIGIT
-        | DIGIT19 ~ DIGIT
-        | DIGIT) ~> (java.lang.Integer.parseInt(_).toByte)
-  }
-
-  def IPv4Address = {
-    rule {
-      IPNumber ~ '.' ~ IPNumber ~ '.' ~ IPNumber ~ '.' ~ IPNumber ~> (Array[Byte](_, _, _, _))
-    }
-  }
-
-  def IPv6Address: Rule1[Array[Byte]] = {
-    import CharUtils.{ hexValue ⇒ hv }
-    var a: Array[Byte] = null
-    def zero(ix: Int) = rule { run(a(ix) = 0.toByte) }
-    def zero2(ix: Int) = rule { run { a(ix) = 0.toByte; a(ix + 1) = 0.toByte; } }
-    def h4(ix: Int) = rule { HEXDIG ~ run(a(ix) = hv(lastChar).toByte) }
-    def h8(ix: Int) = rule { HEXDIG ~ HEXDIG ~ run(a(ix) = (hv(charAt(-2)) * 16 + hv(lastChar)).toByte) }
-    def h16(ix: Int) = rule { h8(ix) ~ h8(ix + 1) | h4(ix) ~ h8(ix + 1) | zero(ix) ~ h8(ix + 1) | zero(ix) ~ h4(ix + 1) }
-    def h16c(ix: Int) = rule { h16(ix) ~ ':' ~ !':' }
-    def ch16o(ix: Int) = rule { optional(':' ~ !':') ~ (h16(ix) | zero2(ix)) }
-    def ls32 = rule { h16(12) ~ ':' ~ h16(14) | IPv4Address ~> (System.arraycopy(_, 0, a, 12, 4)) }
-    def cc(ix: Int) = rule { ':' ~ ':' ~ zero2(ix) }
-    def tail2 = rule { h16c(2) ~ tail4 }
-    def tail4 = rule { h16c(4) ~ tail6 }
-    def tail6 = rule { h16c(6) ~ tail8 }
-    def tail8 = rule { h16c(8) ~ tail10 }
-    def tail10 = rule { h16c(10) ~ ls32 }
-    rule {
-      !(':' ~ HEXDIG) ~ push { a = new Array[Byte](16); a } ~ (
-        h16c(0) ~ tail2
-        | cc(0) ~ tail2
-        | ch16o(0) ~ (
-          cc(2) ~ tail4
-          | ch16o(2) ~ (
-            cc(4) ~ tail6
-            | ch16o(4) ~ (
-              cc(6) ~ tail8
-              | ch16o(6) ~ (
-                cc(8) ~ tail10
-                | ch16o(8) ~ (
-                  cc(10) ~ ls32
-                  | ch16o(10) ~ (
-                    cc(12) ~ h16(14)
-                      | ch16o(12) ~ cc(14))))))))
-    }
-  }
-
-  def IPv6Reference: Rule1[String] = rule { capture("[" ~ oneOrMore(HEXDIG | anyOf(":.")) ~ "]") }
 }
 
 private[http4s] object Rfc2616BasicRules {
