+++ tests/src/test/scala/org/http4s/parser/UriParserSpec.scala
@@ -8,7 +8,7 @@ import org.http4s._
 import scala.util.Success
 import org.http4s.internal.parboiled2._
 
-class IpParser(val input: ParserInput, val charset: NioCharset) extends Parser with Rfc3986Parser {
+class IpParserImpl(val input: ParserInput, val charset: NioCharset) extends Parser with IpParser {
   def CaptureIPv6: Rule1[String] = rule { capture(IpV6Address) }
   def CaptureIPv4: Rule1[String] = rule { capture(IpV4Address) }
 }
@@ -37,14 +37,14 @@ class UriParserSpec extends Http4sSpec {
       } yield (f + "::" + b))
 
       foreach(v) { s =>
-        new IpParser(s, StandardCharsets.UTF_8).CaptureIPv6.run() must be_==(Success(s))
+        new IpParserImpl(s, StandardCharsets.UTF_8).CaptureIPv6.run() must be_==(Success(s))
       }
     }
 
     "parse a IPv4 address" in {
       foreach(0 to 255) { i =>
         val addr = s"$i.$i.$i.$i"
-        new IpParser(addr, StandardCharsets.UTF_8).CaptureIPv4.run() must_==(Success(addr))
+        new IpParserImpl(addr, StandardCharsets.UTF_8).CaptureIPv4.run() must_==(Success(addr))
       }
     }
